User Feedback:
User Feedback: Q: are there any other features or ideas that you would like to see implemented? A: A button to return to the main page.

User Feedback: Q: were there any features that didn't work as expected? A: I struggled to actually start the adventure I had created once I made it. There isn’t anywhere to bring me back to the main menu.

Fix: Added a navigation bar to the top of every page which has a home button so that users can easily navigate the website.